pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtd-dashboard"
version = "0.1.0"
description = "GTD Dashboard CLI for Logseq knowledge graphs"
readme = "README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.10"
authors = [
    {name = "Thomas Bray", email = "thomas.bray@midtown.net"}
]
keywords = ["gtd", "productivity", "logseq", "cli", "dashboard"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
gtd-dashboard = "gtd_dashboard.cli:main"
gtd = "gtd_dashboard.cli:main"

[project.urls]
Homepage = "https://github.com/thomasbray/gtd-dashboard"
Repository = "https://github.com/thomasbray/gtd-dashboard"

[tool.hatch.build.targets.wheel]
packages = ["src/gtd_dashboard"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
